backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-genai==1.69.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
